+++ ghost/core/core/server/services/members/service.js
@@ -4,7 +4,7 @@ const tpl = require('@tryghost/tpl');
 const MembersSSR = require('@tryghost/members-ssr');
 const db = require('../../data/db');
 const MembersConfigProvider = require('./MembersConfigProvider');
-const makeMembersCSVImporter = require('@tryghost/members-importer');
+const makeMembersCSVImporter = require('./importer');
 const MembersStats = require('./stats/MembersStats');
 const memberJobs = require('./jobs');
 const logging = require('@tryghost/logging');
