+++ ghost/core/package.json
@@ -118,7 +118,6 @@
     "@tryghost/member-events": "0.0.0",
     "@tryghost/members-api": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
-    "@tryghost/members-importer": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/members-ssr": "0.0.0",
     "@tryghost/members-stripe-service": "0.0.0",
