+++ ghost/core/test/unit/server/services/members/importer/index.test.js
@@ -1,6 +1,6 @@
 const assert = require('assert/strict');
-const MembersCSVImporter = require('../lib/MembersCSVImporter');
-const makeImporter = require('..');
+const MembersCSVImporter = require('../../../../../../core/server/services/members/importer/MembersCSVImporter');
+const makeImporter = require('../../../../../../core/server/services/members/importer');
 
 describe('makeImporter', function (){
     it('should return an instance of MembersCSVImporter', function (){
