--- ghost/members-importer/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@tryghost/members-importer",
-  "version": "0.0.0",
-  "author": "Ghost Foundation",
-  "private": true,
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "fs-extra": "11.3.0",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/members-csv": "0.0.0",
-    "@tryghost/metrics": "1.0.37",
-    "@tryghost/tpl": "0.1.32",
-    "moment-timezone": "0.5.45"
-  }
-}
